# openapi/second_brain_api.yaml
openapi: "3.0.0"

info:
  title: Second Brain API
  version: "1.0.0"
  description: Personal AI Memory Layer — save and retrieve memories across Claude, ChatGPT, and Gemini.

servers:
  - url: https://your-api-host.railway.app/api/v1

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    MemoryCreate:
      type: object
      required: [content, source, domain]
      properties:
        content:
          type: string
          description: The raw memory text to save.
        source:
          type: string
          enum: [claude, chatgpt, gemini, manual]
        domain:
          type: string
          enum: [career, project, research, learning, personal, meta]
        tags:
          type: array
          items:
            type: string
        importance:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        summary:
          type: string
          description: Short 1-2 sentence summary.
        session_id:
          type: string

    MemoryUpdate:
      type: object
      properties:
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        importance:
          type: integer
          minimum: 1
          maximum: 5
        domain:
          type: string
        summary:
          type: string

    MemoryResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        summary:
          type: string
        source:
          type: string
        domain:
          type: string
        tags:
          type: array
          items:
            type: string
        importance:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        session_id:
          type: string
        similarity:
          type: number
          format: float

paths:
  /memories:
    post:
      operationId: save_memory
      summary: Save a memory
      description: Saves a new memory and automatically generates a vector embedding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryCreate"
      responses:
        "201":
          description: Memory saved successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  created_at:
                    type: string
                    format: date-time
                  status:
                    type: string
                    example: saved

  /memories/search:
    get:
      operationId: search_memories
      summary: Semantic search
      description: Returns the top-K most semantically relevant memories for a natural language query.
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Natural language search query.
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: domain
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: min_importance
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        "200":
          description: List of matching memories ordered by similarity.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryResponse"

  /memories/recent:
    get:
      operationId: get_recent_memories
      summary: Recent memories
      description: Returns the most recent N memories, sorted newest first.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: domain
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of recent memories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryResponse"

  /memories/export:
    get:
      operationId: export_memories
      summary: Export all memories
      description: Exports all memories as a JSON array (no embeddings).
      responses:
        "200":
          description: JSON file download of all memories.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryResponse"

  /memories/{id}:
    patch:
      operationId: update_memory
      summary: Update a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryUpdate"
      responses:
        "200":
          description: Memory updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    example: updated

    delete:
      operationId: delete_memory
      summary: Delete a memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Memory deleted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    example: deleted